refactor: remove underscore from js files
