+++ .gitignore
@@ -1,2 +1,3 @@
 calc-width.xlsx
 ~$calc-width.xlsx
+todo.md
