+++ index.mjs
@@ -1,1 +1,1 @@
-import { container } from "./js/_layout.mjs";
+import { container } from "./js/layout.mjs";
